+++ backend/src/auth/auth.controller.ts
@@ -36,10 +36,9 @@ export class AuthController {
 
   @ApiOperation({ summary: "Is current user authorized" })
   @ApiOkResponse({ status: 200, type: UserModel })
-  @Roles("ADMIN", "USER")
-  @UseGuards(RolesGuard)
   @Get("/me")
   me(@Headers() headers) {
+    console.log(headers);
     return this.authService.me(headers.authorization);
   }